refactor(impl/scripts): reduce duplication (rf2-jkake.15)
Fold the bundle-grep primitives shared by the check-*-bundle scanners
into the existing scripts/lib/read-release-bundle.cjs (the module that
already owns "read the release bundle"):

  - escapeRe / countSubstring were verbatim copies in
    check-bundle-isolation.cjs, check-reagent-slim-bundle-isolation.cjs,
    and check-uix-helix-reagent-free.cjs (3x each).
  - check-reagent-slim-bundle-isolation.cjs carried a full
    re-implementation of readReleaseBlob (its local readAllJs), plus the
    long rf2-z9a06 cljs-runtime/ comment block the lib module already
    documents.
  - check-bundle-isolation.cjs's local countPattern is the same
    pre-built-RegExp count as the new shared countMatches.

The scanners now import escapeRe / countSubstring / countMatches from the
lib. Behaviour-preserving: same output, exit codes, and control flow. The
shared countSubstring/countMatches are null-safe (return 0 on a null
blob); every call site already guards on blob == null before counting, so
no existing path changes.

countOccurrences in check-perf-bundle.cjs is left in place: it is a
single-use helper that joins a pattern list and returns null (not 0) on a
missing blob to signal "count N/A" in the report; folding it would muddy
that semantics for no clarity gain.

The two serve-and-run-*-prod thin wrappers are deliberately
self-contained per-build shims (distinct ROOT/PORT + spec-anchored
docstrings) and are left untouched.

lib public surface: only additive (escapeRe / countMatches /
countSubstring added; readReleaseBlob / listReleaseJsFiles unchanged).
