+++ implementation/scripts/lib/read-release-bundle.cjs
@@ -17,13 +17,26 @@
 // inline and filtered to top-level; this module factors the same fix
 // out so the four sibling scanners share one implementation.
 //
-// Two flavours:
+// Two flavours of reader:
 //   readReleaseBlob(dir)     → concatenated string blob; null when
 //                              `dir` doesn't exist. For substring /
 //                              regex grep contracts.
 //   listReleaseJsFiles(dir)  → array of absolute paths; null when
 //                              `dir` doesn't exist. For per-file work
 //                              (e.g. gzipped-size totalling).
+//
+// Plus the grep primitives the sibling check-*-bundle scanners run
+// against the blob. Each scanner used to carry its own verbatim copies
+// of these (rf2-jkake.15 folded them here so the bundle-grep family
+// shares one implementation):
+//   escapeRe(s)              → escape a string for literal use in a
+//                              RegExp.
+//   countMatches(blob, re)   → number of global matches of a RegExp in
+//                              the blob; 0 when blob is null. Resets the
+//                              RegExp's lastIndex defensively (a /g
+//                              literal carries match state across calls).
+//   countSubstring(blob, s)  → number of occurrences of a literal
+//                              substring (escapeRe + global count).
 
 const fs   = require('fs');
 const path = require('path');
@@ -53,4 +66,29 @@ function readReleaseBlob(dir) {
   return out.join('\n');
 }
 
-module.exports = { readReleaseBlob, listReleaseJsFiles };
+function escapeRe(s) {
+  return s.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+}
+
+function countMatches(blob, re) {
+  if (blob == null) {
+    return 0;
+  }
+  // A caller-supplied /g RegExp literal carries `lastIndex` state across
+  // calls; reset it so the count is independent of prior use.
+  re.lastIndex = 0;
+  const m = blob.match(re);
+  return m ? m.length : 0;
+}
+
+function countSubstring(blob, needle) {
+  return countMatches(blob, new RegExp(escapeRe(needle), 'g'));
+}
+
+module.exports = {
+  readReleaseBlob,
+  listReleaseJsFiles,
+  escapeRe,
+  countMatches,
+  countSubstring,
+};
